requirements.txt
fastapi
uvicorn
python-multipart

torch>=1.0.0
torchvision>=0.2.2
numpy>=1.15.4
tqdm>=4.28.1
pillow>=5.3.0
tensorboardX>=1.6
numpy>=1.24.2